# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Workout Tracker API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invitations:
    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Create an invitation (trainer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationBody"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
  /invitations/{token}:
    get:
      operationId: getInvitation
      tags: [invitations]
      summary: Get invitation details by token (public)
      parameters:
        - $ref: "#/components/parameters/InvitationToken"
      responses:
        "200":
          description: Invitation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationDetail"
        "404":
          description: Not found or expired
  /invitations/{token}/accept:
    post:
      operationId: acceptInvitation
      tags: [invitations]
      summary: Accept an invitation and set credentials
      parameters:
        - $ref: "#/components/parameters/InvitationToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitationBody"
      responses:
        "200":
          description: Invitation accepted and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "404":
          description: Not found or expired
        "409":
          description: Username or email already taken
...
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List workout plans
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
          description: Filter plans assigned to this user (or global plans)
      responses:
        "200":
          description: List of plans (without nested sets)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanSummary"
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create a workout plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanBody"
      responses:
        "201":
          description: Created plan with full structure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanDetail"
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get a plan with full structure
      parameters:
        - $ref: "#/components/parameters/PlanId"
      responses:
        "200":
          description: Plan with sets and exercises
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanDetail"
        "404":
          description: Not found
    put:
      operationId: updatePlan
      tags: [plans]
      summary: Update a workout plan (replaces all sets)
      parameters:
        - $ref: "#/components/parameters/PlanId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanBody"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanDetail"
        "404":
          description: Not found
    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete a workout plan
      parameters:
        - $ref: "#/components/parameters/PlanId"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
...
    PlanId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    InvitationToken:
      name: token
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        role:
          type: string
          enum: [trainer, client]
        organizationId:
          type: integer
        organizationName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - organizationId
        - organizationName
        - createdAt
    CreateInvitationBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [trainer, client]
      required:
        - name
        - email
        - role
    InvitationResponse:
      type: object
      properties:
        id:
          type: integer
        token:
          type: string
        inviteUrl:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [trainer, client]
        expiresAt:
          type: string
          format: date-time
      required:
        - id
        - token
        - inviteUrl
        - name
        - email
        - role
        - expiresAt
    InvitationDetail:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [trainer, client]
        organizationName:
          type: string
        expiresAt:
          type: string
          format: date-time
        expired:
          type: boolean
        accepted:
          type: boolean
      required:
        - name
        - email
        - role
        - organizationName
        - expiresAt
        - expired
        - accepted
    AcceptInvitationBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
...
    PlanSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdBy:
          type: integer
          nullable: true
        isGlobal:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        setCount:
          type: integer
        assignedUserIds:
          type: array
          items:
            type: integer
      required:
        - id
        - name
        - isGlobal
        - createdAt
        - updatedAt
        - setCount
        - assignedUserIds
    SetExerciseInput:
      type: object
      properties:
        exerciseId:
          type: integer
        targetValue:
          type: string
        orderIndex:
          type: integer
      required:
        - exerciseId
        - targetValue
        - orderIndex
    PlanSetInput:
      type: object
      properties:
        type:
          type: string
          enum: [straight, superset, triset, other]
        rounds:
          type: integer
        restSeconds:
          type: integer
          nullable: true
        orderIndex:
          type: integer
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/SetExerciseInput"
      required:
        - type
        - rounds
        - orderIndex
        - exercises
    CreatePlanBody:
      type: object
      properties:
        name:
          type: string
        isGlobal:
          type: boolean
        assignedUserIds:
          type: array
          items:
            type: integer
        createdBy:
          type: integer
        sets:
          type: array
          items:
            $ref: "#/components/schemas/PlanSetInput"
      required:
        - name
        - sets
    UpdatePlanBody:
      type: object
      properties:
        name:
          type: string
        isGlobal:
          type: boolean
        assignedUserIds:
          type: array
          items:
            type: integer
        sets:
          type: array
          items:
            $ref: "#/components/schemas/PlanSetInput"
      required:
        - name
        - sets
    SetExerciseDetail:
      type: object
      properties:
        id:
          type: integer
        exerciseId:
          type: integer
        exerciseName:
          type: string
        exerciseMeasurementType:
          type: string
          enum: [reps, seconds, meters]
        targetValue:
          type: string
        orderIndex:
          type: integer
      required:
        - id
        - exerciseId
        - exerciseName
        - exerciseMeasurementType
        - targetValue
        - orderIndex
    PlanSetDetail:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [straight, superset, triset, other]
        rounds:
          type: integer
        restSeconds:
          type: integer
          nullable: true
        orderIndex:
          type: integer
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/SetExerciseDetail"
      required:
        - id
        - type
        - rounds
        - orderIndex
        - exercises
    PlanDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdBy:
          type: integer
          nullable: true
        isGlobal:
          type: boolean
        assignedUserIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sets:
          type: array
          items:
            $ref: "#/components/schemas/PlanSetDetail"
      required:
        - id
        - name
        - isGlobal
        - assignedUserIds
        - createdAt
        - updatedAt
        - sets
...